+++ bld/wgml2/c/gsli.c
@@ -105,7 +105,7 @@ void    scr_li( void )
             case 2 :
                 if( !strnicmp( "ON", pa, 2 ) ) {
                     if( !ProcFlags.literal ) {
-                        li_cnt = LONG_MAX;
+                        li_cnt = INT_MAX;
                         ProcFlags.literal = true;
                         scan_restart = pa + 2;
                     }
