+++ bld/wgml2/c/wgml.c
@@ -401,7 +401,7 @@ static  bool    test_comment( void )
                 (*(buff2 + 4) == '.')  ) {
 
                 if( ProcFlags.literal ) {   // special
-                    if( li_cnt < LONG_MAX ) {// we decrement, do not wait for .li OFF
+                    if( li_cnt < INT_MAX ) {// we decrement, do not wait for .li OFF
                         if( li_cnt-- <= 0 ) {
                             ProcFlags.literal = false;
                         }
